Update chapter-12.md
The guide on stackoverflow is heavily outdated and is missing critical steps. I have added the steps here. I've also added a line which asks us to ensure datapool mounts on boot.

A lot of people had issues with lab-12 and have messaged me about it. These instructions should help people in the future.